# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Approval Workflow API",
        "description": "REST API Approval Workflow",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/approval-requests/{id}/submit": {
            "post": {
                "tags": [
                    "Approval Workflow"
                ],
                "summary": "Submit Request",
                "description": "Employee submit draft request menjadi submitted.",
                "operationId": "a0ab3c1253762f7fd45af2e200b198a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request berhasil disubmit"
                    },
                    "400": {
                        "description": "Request sudah diproses"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/approval-requests/{id}/approve": {
            "post": {
                "tags": [
                    "Approval Workflow"
                ],
                "summary": "Approve Request",
                "description": "Manager menyetujui request.",
                "operationId": "1896bc37f8271ddbd775d70e9d6d1f38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment"
                                ],
                                "properties": {
                                    "comment": {
                                        "type": "string",
                                        "example": "Request disetujui."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request berhasil diapprove"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/approval-requests/{id}/reject": {
            "post": {
                "tags": [
                    "Approval Workflow"
                ],
                "summary": "Reject Request",
                "description": "Manager menolak request.",
                "operationId": "15c3ee1e2e7b6a5aba24795b13e9a394",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment"
                                ],
                                "properties": {
                                    "comment": {
                                        "type": "string",
                                        "example": "Budget belum tersedia."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request berhasil direject"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/approval-requests/{id}/history": {
            "get": {
                "tags": [
                    "Approval History"
                ],
                "summary": "Approval History",
                "description": "Menampilkan seluruh riwayat approval request.",
                "operationId": "af6a4b90486053dcbabac9d76f3992eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "History berhasil diambil"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/approval-requests/{id}/timeline": {
            "get": {
                "tags": [
                    "Approval History"
                ],
                "summary": "Approval Timeline",
                "description": "Menampilkan timeline perubahan status approval request.",
                "operationId": "ee244db6c48a80415d92e23057fb171e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Timeline berhasil diambil"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Masukkan token dengan format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}